+++ .ci/changelog.jl
@@ -4,6 +4,6 @@ cd(dirname(@__DIR__)) do
     Changelog.generate(
         Changelog.CommonMark(),
         "CHANGELOG.md";
-        repo="MichaelHatherly/HypertextTemplates.jl",
+        repo = "MichaelHatherly/HypertextTemplates.jl",
     )
 end